+++ CHANGELOG.md
@@ -69,6 +69,8 @@ Remember to align the itemized text with the first line of an item within a list
     {func}`~jax.numpy.apply_over_axes`, {func}`~jax.numpy.inner`,
     {func}`~jax.numpy.outer`, {func}`~jax.numpy.cross`,
     {func}`~jax.numpy.kron`, and {func}`~jax.numpy.lexsort`.
+  * The ``kind`` argument to {func}`jax.numpy.sort` and {func}`jax.numpy.argsort`
+    is now removed. Use `stable=True` or `stable=False` instead.
 
 * Bug fixes
   * {func}`jax.numpy.astype` will now always return a copy when `copy=True`.